Added storage size check variables
Added necessary storage array size variables to private, and also to constructor.  Supported size values are "4" and "16", corresponding to the 47*04 and 47*16 parts respectively.  This option will help prevent erroneous reads from undesired addressed if the called address falls outised the actual size of the EERAM array.  The problem here is that while the address is necessarily a 16bit int, the chip only checks the least significant 11 bits.  So for example if a user were to ask for data from or try to write data to "0xE5B5", the chip would happily apply that operation to the address "0x5B5", which would likely give undesirable results.  The size of the array in the constructor defaults to 16kbit so users that don't specify the option don't encounter errors when using a 16kbit part (which seems more popular that the 4kbit part anyways).